config  entity
