+++ src/main/java/com/healthy/BackEnd/BackEndApplication.java
@@ -1,0 +1,13 @@
+package com.healthy.BackEnd;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BackEndApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(BackEndApplication.class, args);
+	}
+
+}
