+++ src/main/java/com/healthy/BackEnd/Controller/UserController.java
@@ -1,0 +1,37 @@
+package com.healthy.BackEnd.Controller;
+
+
+import com.healthy.BackEnd.Service.AuthenticationService;
+import com.healthy.BackEnd.dto.LoginDTO;
+import com.healthy.BackEnd.entity.Users;
+import jakarta.validation.Valid;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("api")
+public class UserController {
+    @Autowired
+    AuthenticationService authenticationService;
+
+    @PostMapping("/register")
+    public ResponseEntity<?> register (@Valid @RequestBody Users users) {
+        Users newAccount = authenticationService.register(users);
+        return ResponseEntity.ok(newAccount);
+    }
+
+    @PostMapping("/login")
+    public ResponseEntity<?> loginEmployee(@Valid @RequestBody LoginDTO loginDTO) {
+        boolean isAuthenticated = authenticationService.login(loginDTO);
+
+        if (isAuthenticated) {
+            return ResponseEntity.ok("Login successful");
+        } else {
+            return ResponseEntity.status(401).body("Invalid username or password");
+        }
+    }
+}
