+++ src/main/java/com/healthy/BackEnd/dto/LoginDTO.java
@@ -1,0 +1,15 @@
+package com.healthy.BackEnd.dto;
+
+import jakarta.validation.constraints.NotBlank;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class LoginDTO {
+    @NotBlank(message = "Username is required")
+    private String username;
+
+    @NotBlank(message = "Password is required")
+    private String password;
+}
