+++ src/main/java/com/healthy/BackEnd/entity/Students.java
@@ -1,0 +1,42 @@
+package com.healthy.BackEnd.entity;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.Setter;
+
+@Entity
+@Getter
+@Setter
+@Table(name = "Students")
+public class Students {
+    
+    @Id
+    @Column(name = "StudentID", length = 36, nullable = false)
+    private String studentID;
+
+    @Column(name = "UserID", length = 36, nullable = false)
+    private String userID;
+
+    @Column(name = "Grade")
+    private Integer grade;
+
+    @Column(name = "Class", length = 20)
+    private String className;
+
+    @Column(name = "SchoolName", length = 100)
+    private String schoolName;
+
+    @Enumerated(EnumType.STRING)
+    @Column(name = "Gender", columnDefinition = "ENUM('Male', 'Female', 'Other')")
+    private Gender gender;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "UserID", referencedColumnName = "UserID", insertable = false, updatable = false)
+    private Users user;
+
+    public enum Gender {
+        Male,
+        Female,
+        Other
+    }
+}
