+++ src/main/java/com/healthy/BackEnd/repository/AuthenticationRepo.java
@@ -1,0 +1,10 @@
+package com.healthy.BackEnd.repository;
+
+import com.healthy.BackEnd.entity.Users;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AuthenticationRepo extends JpaRepository<Users, String> {
+    Users findByUsername(String username);
+
+    }
+
