--- src/main/java/com/heathly/BackEnd/BackEndApplication.java
@@ -1,13 +1,0 @@
-package com.heathly.BackEnd;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class BackEndApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(BackEndApplication.class, args);
-	}
-
-}
