--- src/main/java/com/heathly/BackEnd/MyNamingStrategy.java
@@ -1,45 +1,0 @@
-package com.heathly.BackEnd;
-
-import org.hibernate.boot.model.naming.Identifier;
-import org.hibernate.boot.model.naming.PhysicalNamingStrategy;
-import org.hibernate.engine.jdbc.env.spi.JdbcEnvironment;
-
-public class MyNamingStrategy implements PhysicalNamingStrategy {
-    @Override
-    public Identifier toPhysicalCatalogName(Identifier identifier, JdbcEnvironment jdbcEnvironment) {
-        return null;
-    }
-
-    @Override
-    public Identifier toPhysicalSchemaName(Identifier identifier, JdbcEnvironment jdbcEnvironment) {
-        return null;
-    }
-
-    @Override
-    public Identifier toPhysicalTableName(Identifier name, JdbcEnvironment context) {
-        String tableName = convertToPascalCase(name.getText());
-        return new Identifier(tableName, name.isQuoted());
-    }
-
-    @Override
-    public Identifier toPhysicalSequenceName(Identifier identifier, JdbcEnvironment jdbcEnvironment) {
-        return null;
-    }
-
-    @Override
-    public Identifier toPhysicalColumnName(Identifier name, JdbcEnvironment context) {
-        String columnName = convertToPascalCase(name.getText());
-        return new Identifier(columnName, name.isQuoted());
-    }
-
-    private String convertToPascalCase(String text) {
-        StringBuilder result = new StringBuilder();
-        for (String part : text.split("_")) {
-            if (!part.isEmpty()) {
-                result.append(part.substring(0, 1).toUpperCase());
-                result.append(part.substring(1).toLowerCase());
-            }
-        }
-        return result.toString();
-    }
-}
