--- src/main/java/com/heathly/BackEnd/dto/LoginDTO.java
@@ -1,15 +1,0 @@
-package com.heathly.BackEnd.dto;
-
-import jakarta.validation.constraints.NotBlank;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class LoginDTO {
-    @NotBlank(message = "Username is required")
-    private String username;
-
-    @NotBlank(message = "Password is required")
-    private String password;
-}
