--- src/main/java/com/heathly/BackEnd/entity/TimeSlots.java
@@ -1,69 +1,0 @@
-package com.heathly.BackEnd.entity;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.LocalTime;
-
-@Entity
-@Getter
-@Setter
-@Table(name = "TimeSlots")
-public class TimeSlots {
-    
-    @Id
-    @Column(name = "TimeSlotsID", length = 36, nullable = false)
-    private String timeSlotsID;
-
-    @Column(name = "PsychologistID", length = 36, nullable = false)
-    private String psychologistID;
-
-    @Column(name = "SlotDate", nullable = false)
-    private LocalDate slotDate;
-
-    @Column(name = "StartTime", nullable = false)
-    private LocalTime startTime;
-
-    @Column(name = "EndTime", nullable = false)
-    private LocalTime endTime;
-
-    @Enumerated(EnumType.STRING)
-    @Column(name = "Status", columnDefinition = "ENUM('Available', 'Booked')", nullable = false)
-    private Status status;
-
-    @Column(name = "CreatedAt", nullable = false, updatable = false)
-    private LocalDateTime createdAt;
-
-    @Column(name = "UpdatedAt", nullable = false)
-    private LocalDateTime updatedAt;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "PsychologistID", referencedColumnName = "PsychologistID", insertable = false, updatable = false)
-    private Psychologists psychologist;
-
-    public TimeSlots() {
-        status = Status.Available;
-    }
-
-    @PrePersist
-    protected void onCreate() {
-        if (createdAt == null) {
-            createdAt = LocalDateTime.now();
-        }
-        if (updatedAt == null) {
-            updatedAt = LocalDateTime.now();
-        }
-    }
-
-    @PreUpdate
-    protected void onUpdate() {
-        updatedAt = LocalDateTime.now();
-    }
-
-    public enum Status {
-        Available,
-        Booked
-    }
-}
