+++ src/main/resources/application.properties
@@ -7,7 +7,7 @@ spring.datasource.password=12345678
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL8Dialect
 spring.jpa.hibernate.ddl-auto=update
 spring.main.allow-circular-references=true
-hibernate.physical_naming_strategy=com.heathly.BackEnd.MyNamingStrategy
+hibernate.physical_naming_strategy=com.healthy.BackEnd.PascalCaseNamingStrategy
 spring.jpa.hibernate.naming.physical-strategy = org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
 spring.jpa.hibernate.naming.implicit-strategy = org.hibernate.boot.model.naming.ImplicitNamingStrategyLegacyJpaImpl
