+++ src/test/java/com/healthy/BackEnd/BackEndApplicationTests.java
@@ -1,0 +1,13 @@
+package com.healthy.BackEnd;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BackEndApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
