--- src/test/java/com/heathly/BackEnd/BackEndApplicationTests.java
@@ -1,13 +1,0 @@
-package com.heathly.BackEnd;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class BackEndApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
